pyproject.toml
[project]
name = "ai-coach"
version = "0.1.0"
description = "AI-assisted candidate profile ingestion pipeline"
authors = [{name = "AI Coach"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.16",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "pypdf>=4.2.0",
    "python-docx>=1.1.0",
    "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
